chore(electron): batch A — real fixes for 6 small electron lint errors (#223)
Three concrete refactors, no rule suppressions.

1. pdf.tsx:755 — drop `async` from `applyVisual` (require-await).
   `HighlightTarget.applyVisual` is already typed as `() => Promise<void> | void`
   (highlight-actions.ts:18), so the async marker was just lint noise.
   The function body has no await; removing async returns void, which the
   interface accepts directly. Caller still `await args.target.applyVisual()`
   — awaiting void resolves synchronously.

2. local-vad.test.ts — replace empty connect/disconnect methods with
   vi.fn() arrow properties (no-empty-function x4). FakeAnalyser and
   FakeStreamSource fake the Web Audio API for happy-dom tests. Empty
   methods satisfied the interface but yielded zero observability —
   any future test that wanted to assert "we connected then disconnected"
   would have had to monkey-patch. vi.fn() satisfies the interface
   contract AND makes the call sites spyable, with no behavioural change
   to existing tests (none of which observe these calls today).

3. pdf-page.tsx — extract `makeCustomTextRenderer` to its own file
   `makeCustomTextRenderer.ts` (react-refresh/only-export-components).
   The function was already documented as "extracted from the component
   closure so the declarative TTS-highlight behavior can be unit-tested"
   — the docstring even pointed at this — so the function belongs in
   its own module. Now Fast Refresh works for PageComponent without
   tripping the HMR boundary rule. Updated the test import path too.

Lint state:
  before:  ✖ 38 problems (38 errors)
  after:   ✖ 69 problems (32 errors, 37 warnings)
                ^^^^^^^^^^^^^^^^ -6 errors

Typecheck (both tsc projects) passes clean. No semantic regressions.

Remaining 32 errors tracked in #223:
  23  no-unnecessary-condition  (per-site type narrowing)
   7  react-hooks/refs          (real refactor — ref access during render)
   2  set-state-in-effect       (real refactor)
